+++ happn-cluster/lib/services/replicator.js
@@ -160,6 +160,10 @@ Replicator.prototype.unbatchSecurityUpdate = function (payload) {
   unbatched = unbatched.concat(this.unpackBywhatHappnd(payload, SD_EVENTS.PERMISSION_UPSERTED));
   unbatched = unbatched.concat(this.unpackBywhatHappnd(payload, SD_EVENTS.DELETE_USER));
   unbatched = unbatched.concat(this.unpackBywhatHappnd(payload, SD_EVENTS.DELETE_GROUP));
+  unbatched = unbatched.concat(this.unpackBywhatHappnd(payload, SD_EVENTS.LOOKUP_TABLE_CHANGED));
+  unbatched = unbatched.concat(
+    this.unpackBywhatHappnd(payload, SD_EVENTS.LOOKUP_PERMISSION_CHANGED)
+  );
   return unbatched;
 };