+++ happn-cluster/test/test-unit/05-unit-replicator.js
@@ -364,6 +364,96 @@ require('../lib/test-helper').describe({ timeout: 30e3 }, function (test) {
       .to.be('unknown security data changed event: unknown');
   });
 
+  it('tests the unbatchSecurityUpdate method', function (done) {
+    const replicator = new Replicator(mockOpts);
+    replicator.happn = new MockHappn('http', 9000);
+    let payload = {};
+
+    payload[SD_EVENTS.LINK_GROUP] = {
+      key1: { changedData: 'changed1', additionalInfo: 'additional1' },
+    };
+    payload[SD_EVENTS.UNLINK_GROUP] = {
+      key2: { changedData: 'changed2', additionalInfo: 'additional2' },
+    };
+    payload[SD_EVENTS.UPSERT_GROUP] = {
+      key3: { changedData: 'changed3', additionalInfo: 'additional3' },
+    };
+    payload[SD_EVENTS.UPSERT_USER] = {
+      key4: { changedData: 'changed4', additionalInfo: 'additional4' },
+    };
+    payload[SD_EVENTS.PERMISSION_REMOVED] = {
+      key5: { changedData: 'changed5', additionalInfo: 'additional5' },
+    };
+    payload[SD_EVENTS.PERMISSION_UPSERTED] = {
+      key6: { changedData: 'changed6', additionalInfo: 'additional6' },
+    };
+    payload[SD_EVENTS.DELETE_USER] = {
+      key7: { changedData: 'changed7', additionalInfo: 'additional7' },
+    };
+    payload[SD_EVENTS.DELETE_GROUP] = {
+      key8: { changedData: 'changed8', additionalInfo: 'additional8' },
+    };
+    payload[SD_EVENTS.LOOKUP_TABLE_CHANGED] = {
+      key9: { changedData: 'changed9', additionalInfo: 'additional9' },
+    };
+    payload[SD_EVENTS.LOOKUP_PERMISSION_CHANGED] = {
+      key10: { changedData: 'changed10', additionalInfo: 'additional10' },
+    };
+    test.expect(replicator.unbatchSecurityUpdate(payload)).to.eql([
+      {
+        whatHappnd: SD_EVENTS.LINK_GROUP,
+        changedData: 'changed1',
+        additionalInfo: 'additional1',
+      },
+      {
+        whatHappnd: SD_EVENTS.UNLINK_GROUP,
+        changedData: 'changed2',
+        additionalInfo: 'additional2',
+      },
+      {
+        whatHappnd: SD_EVENTS.UPSERT_GROUP,
+        changedData: 'changed3',
+        additionalInfo: 'additional3',
+      },
+      {
+        whatHappnd: SD_EVENTS.UPSERT_USER,
+        changedData: 'changed4',
+        additionalInfo: 'additional4',
+      },
+      {
+        whatHappnd: SD_EVENTS.PERMISSION_REMOVED,
+        changedData: 'changed5',
+        additionalInfo: 'additional5',
+      },
+      {
+        whatHappnd: SD_EVENTS.PERMISSION_UPSERTED,
+        changedData: 'changed6',
+        additionalInfo: 'additional6',
+      },
+      {
+        whatHappnd: SD_EVENTS.DELETE_USER,
+        changedData: 'changed7',
+        additionalInfo: 'additional7',
+      },
+      {
+        whatHappnd: SD_EVENTS.DELETE_GROUP,
+        changedData: 'changed8',
+        additionalInfo: 'additional8',
+      },
+      {
+        whatHappnd: SD_EVENTS.LOOKUP_TABLE_CHANGED,
+        changedData: 'changed9',
+        additionalInfo: 'additional9',
+      },
+      {
+        whatHappnd: SD_EVENTS.LOOKUP_PERMISSION_CHANGED,
+        changedData: 'changed10',
+        additionalInfo: 'additional10',
+      },
+    ]);
+    done();
+  });
+
   it('can call the send function, security update - emit', function (done) {
     const replicator = new Replicator(mockOpts);
     const emitted = [];